requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
boto3==1.26.147
botocore==1.29.147
cffi==1.15.1
cryptography==41.0.1
decorator==5.1.1
Django==4.2.2
django-filter==23.2
django-formset==0.13.4
django-material==1.12.0
django-viewflow==2.0.0b4
fabric2==3.1.0
gunicorn==20.1.0
invoke==2.1.3
jmespath==1.0.1
multi-sqs-listener==1.0.1
paramiko==3.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
